requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.9.0
python-dotenv==1.0.1
streamlit==1.39.0
google-genai==1.42.0
sqlite-utils==3.36
SQLAlchemy==2.0.32
requests==2.32.3
python-multipart==0.0.9
httpx==0.27.0